[msbuild/dotnet] Rework how we handle debug symbols and executable stripping. Fixes #14067.
----

This turned out somewhat non-straight-forward for universal builds, because debug
symbols are created for the universal app bundle, not for each rid-specific version
of the app bundle.

So I had to add logic to create the native symbol lists (MtouchSymbolsList) for each
rid-specific build, but then collect them and merge those lists for the universal
app bundle.

The existing SymbolStrip call we did right after linking the native executable was
removed, because we have to do that after creating the dSYM (which the GenerateDebugSymbols
target does).

Also add tests.

---

Fixes https://github.com/xamarin/xamarin-macios/issues/14067.