[dotnet] Use only the first two parts of our product version + the version band as the assembly version.
This is a follow up to 92ee92eeb57f470eea28a2119f7d329be6a24030, where we
changed the assembly version from four digits to three digits (i.e not include
the build number), because our managed API should be identical for (released)
versions where only the build number changes.

In this PR we go a bit further, because we're not supposed to have any API
differences unless the version band changes, which means that we should only
use need the two first parts of our product version + the version band as the
assembly version.