[msbuild] Don't add 'native' assets to the hot restart content directory.
NuGet adds 'native' assets without a directory component (see
https://github.com/dotnet/sdk/issues/9643), which means that for frameworks in
NuGets all files in those frameworks are added as plain files, without any
information about which subdirectory to put them in. This is obviously wrong,
so just don't add any native assets to the content directory.

Additionally, frameworks shouldn't be in the hot restart content directory in the first place, because:

* They need to be signed in order to execute on the device.
* iOS won't allow executing native code from a writeable directory
* As such they can't be copied on demand, they have to be installed with the app.