[msbuild] Fix the target path of a compressed plugin.
A compressed file representing a plugin is extracted directly into the [Contents]/PlugIns
directory, and the name of the compressed file is not taken into account.

Example zip, containing a MyPlugin.plugin directory containing a MyFile.data:

    MyPlugin.zip
        MyPlugin.plugin
			MyFile.data

Would end up like this on iOS:

    MyApp.app
        PlugIns
            MyPlugin.plugin
                MyFile.data

This means that we must not take the zip file's name into accoun when computing the
bundle location.