[msbuild] Create a temporary directory to host the temporary project for the FindAotCompiler task.
This speeds up the FindAotCompiler task tremendously when /tmp is full of
stuff: .NET projects will implicity include a lot of files in the project
directory, and if we put the project file in /tmp, then MSBuild will search
all of /tmp for files to include (and include some of them). This is avoided
by putting the project file in a subdirectory with nothing else in it.

Also add the binlogs and the output text file in the new temporary directory,
which means we can simplify the cleanup code as well (to just remove the
temporary directory, instead of each individual file).