[dotnet] Fix entitlements for the createdump executable. Fixes #16189.
We can't use the entitlements for the main executable as-is for the createdump
executable (that brings in entitlements that are only valid for the main
executable) so implement logic to compile entitlements for createdump
separately, and only include a very specific list of entitlements (ones that
are listed as required by .NET + one that the App Store complains about if
it's not there).

This also means adding support to the CompileEntitlements task to not embed any default
entitlements or entitlements from the provisioning profile.

Also add a property to opt out of bundling createdump in the app bundle (`BundleCreateDump`).

Fixes https://github.com/xamarin/xamarin-macios/issues/16189.