[tests] Zip up the app bundle we built and inspect the zipped file on Windows.
Zip up the app bundle we built in the BundleStructure test and instead of verifying
the build results we end up with in the project directory on Windows, we instead
verify the contents of the zipped app bundle.

This is because the app bundle the build produces on Windows isn't a working app
bundle, many files aren't there, and other files are just a 0-length stamp file to
help with incremental builds (this is by design to not have to copy a lot of data
back from the Mac to the Windows machine in order to speed up the build).