[dotnet] Make the VB templates compile.
VB doesn't support default interface members, and will show an error for all unimplemented
interface members, even if they have a default implementation.

This is a problem for the VB templates, because they make use of protocols, which
will soon have interface members with default implementations.

In order to get around this, change the VB templates to use the model class instead
of the protocol interfaces.