[msbuild] Show an error if an app tries to link with a static library when building for Hot Restart.
A fairly common problem is when apps try to use a static library when using
Hot Restart on Windows (which isn't supported), and then they're confused when
their app doesn't work.

Improve this scenario by raising an error if an app tries to link with a
static library when building for Hot Restart.